refactor(core): improve db module and cleanup codebase
- Extract common query utilities (isSQLWrapper, buildWhereFromObject) to query-utils.ts
- Optimize count() to use SQL COUNT instead of fetching all rows
- Integrate env module parsers in db/manager/config.ts
- Remove unused imports and fix type errors in tests
- Add env module tests (registry, schema, validator)
- Add error-handler middleware test
- Update docs meta files

+++ packages/core/src/env/index.ts
@@ -33,7 +33,7 @@ export {
     parseNumber,
     createNumberParser,
     parseInteger,
-    parseFloat,
+    parseDecimal,
     // URL parsers
     parseUrl,
     createUrlParser,
